dashboard/requirements.txt
flask
matplotlib
pandas
requests
python-jenkins
openpyxl
